moved fetchContactsData to contacts/index
