+++ app/javascript/src/components/Dashboard/Contacts/Table/index.jsx
@@ -1,15 +1,14 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect } from "react";
 
 import { Table as NeetoUITable } from "neetoui";
 import { noop } from "utils";
 
 import Dropdown from "./Dropdown";
 import NameRoleAndAvatar from "./NameRoleAndAvatar";
-import { buildContactTableColumns, getContactsData } from "./utils";
 
-const Table = () => {
-  const [contactsData, setContactsData] = useState([]);
+import { buildContactTableColumns } from "../utils";
 
+const Table = ({ fetchContactsData, contactsData }) => {
   const renderNameAndRole = (nameAndRole, { name, role }) => (
     <NameRoleAndAvatar name={name} role={role} />
   );
@@ -19,11 +18,6 @@ const Table = () => {
     renderNameAndRole,
     renderDropdown
   );
-
-  const fetchContactsData = () => {
-    // TODO: Get data through API, once implemented
-    setContactsData(getContactsData());
-  };
 
   useEffect(() => {
     fetchContactsData();
